A upgrade to `v1.10.0` with a punishement for the crackers

+++ Panel/build.js
@@ -1,24 +1,35 @@
 const fs = require("fs");
 const path = require("path");
-const { execSync } = require("child_process");
+const { exec } = require("child_process");
 
-let resellers_data = require("../reseller.config.json"); const
-    { currentBuildFor } = resellers_data; delete resellers_data.currentBuildFor;
+const resellers_data = require("../reseller.config.json"),
+    resellers_starting_data = fs.readFileSync(path.join(__dirname, "..", "reseller.config.json"));
 
-
-Object.keys(resellers_data).forEach(async reseller => {
+delete resellers_data.currentBuildFor;
+const sellers = Object.keys(resellers_data); (async function startSellers(i) {
+    const reseller = sellers[i];
     fs.writeFileSync(path.join(__dirname, "..", "reseller.config.json"), JSON.stringify({
         ...resellers_data, ...{
             currentBuildFor: reseller
         }
     }));
 
-    const log = execSync(`npm run build:tailwindcss && node build.asar.js && pkg -C GZip -o dist/${reseller}.exe .`); if (fs.existsSync(path.join(__dirname, "dist", `${reseller}.exe`))) {
-        fs.copyFileSync(path.join(__dirname, "dist", `${reseller}.exe`), path.join(__dirname, "dist_exe", `${reseller}.exe`));
+    const proc = exec(`npm run build:tailwindcss && node build.asar.js --build && cd dist && pkg -o ${reseller}.exe .`);
 
-        console.log(`Build for ${reseller} done.`);
-    } else
-        console.log(`Failed Build for ${reseller}`, log);
-});
+    proc.stdout.pipe(process.stdout);
+    proc.stderr.pipe(process.stderr);
+    proc.addListener("exit", (code) => {
+        if (fs.existsSync(path.join(__dirname, "dist", `${reseller}.exe`)) && code == 0) {
+            fs.copyFileSync(path.join(__dirname, "dist", `${reseller}.exe`), path.join(__dirname, "dist_exe", `${reseller}.exe`));
 
-fs.writeFileSync(path.join(__dirname, "..", "reseller.config.json"), JSON.stringify({ ...resellers_data, ...{ currentBuildFor } }));
+            console.log(`Build for ${reseller} done.`);
+        } else
+            console.log(`Failed Build for ${reseller}`);
+
+        if ((i + 1) < sellers.length)
+            startSellers(i++);
+        else {
+            fs.writeFileSync(path.join(__dirname, "..", "reseller.config.json"), resellers_starting_data);
+        };
+    });
+})(0);
